// generated-tokens/tokens.ts
// Auto-generated by Style Dictionary
import { colorPrimitives } from "./colorPrimitives";
export const tokens = {
  layer: {
    background: colorPrimitives.gray["000"],
    primary: colorPrimitives.grayAlpha["100"],
    secondary: colorPrimitives.gray["000"],
    secondaryPressed: colorPrimitives.grayAlpha["100"],
    backgroundPressed: colorPrimitives.grayAlpha["100"],
    primaryPressed: colorPrimitives.grayAlpha["100"],
  },
  object: {
    primary: {
      bold: {
        default: colorPrimitives.yellow["300"],
        pressed: colorPrimitives.yellow["400"],
        selected: colorPrimitives.yellow["200"],
      },
      subtle: {
        default: colorPrimitives.yellow["100"],
        pressed: colorPrimitives.yellow["300"],
        selected: colorPrimitives.yellow["300"],
      },
    },
    secondary: {
      default: colorPrimitives.gray["100"],
      pressed: colorPrimitives.gray["200"],
      selected: colorPrimitives.yellow["300"],
    },
    tertiary: {
      default: colorPrimitives.gray["000"],
      pressed: colorPrimitives.grayAlpha["100"],
      selected: colorPrimitives.yellow["300"],
    },
    negative: {
      subtle: {
        default: colorPrimitives.red["100"],
        pressed: colorPrimitives.red["300"],
        selected: colorPrimitives.red["600"],
      },
      bold: {
        default: colorPrimitives.red["600"],
        pressed: colorPrimitives.red["700"],
        selected: colorPrimitives.red["400"],
      },
    },
    positive: {
      subtle: {
        default: colorPrimitives.green["100"],
        pressed: colorPrimitives.green["300"],
        selected: colorPrimitives.green["500"],
      },
      bold: {
        default: colorPrimitives.green["500"],
        pressed: colorPrimitives.green["600"],
        selected: colorPrimitives.green["300"],
      },
    },
    inverse: {
      default: colorPrimitives.gray["900"],
      pressed: colorPrimitives.gray["800"],
      selected: colorPrimitives.gray["700"],
    },
    disabled: {
      disabled: colorPrimitives.grayAlpha["300"],
    },
    accent: {
      subtle: {
        selected: colorPrimitives.blue["600"],
        pressed: colorPrimitives.blue["300"],
        default: colorPrimitives.blue["000"],
      },
      bold: {
        selected: colorPrimitives.blue["400"],
        pressed: colorPrimitives.blue["700"],
        default: colorPrimitives.blue["600"],
      },
    },
  },
  face: {
    primary: colorPrimitives.gray["900"],
    secondary: colorPrimitives.gray["800"],
    tertiary: colorPrimitives.gray["700"],
    negativeBold: colorPrimitives.red["700"],
    positiveBold: colorPrimitives.green["700"],
    disabled: colorPrimitives.gray["400"],
    inversePrimary: colorPrimitives.gray["000"],
    accentBold: colorPrimitives.blue["700"],
    positiveSubtle: colorPrimitives.green["000"],
    negativeSubtle: colorPrimitives.red["000"],
    accentSubtle: colorPrimitives.blue["000"],
    inverseSecondary: colorPrimitives.gray["200"],
    inverseTertiary: colorPrimitives.gray["300"],
  },
  border: {
    secondary: colorPrimitives.gray["200"],
    negative: colorPrimitives.red["700"],
    primary: colorPrimitives.yellow["400"],
    positive: colorPrimitives.green["700"],
    disabled: colorPrimitives.grayAlpha["200"],
    focused: colorPrimitives.blue["700"],
    inverse: colorPrimitives.gray["000"],
    tertiary: colorPrimitives.gray["100"],
  },
  spacing: {
    xs: 6,
    md: 12,
    lg: 16,
    sm: 8,
    xl: 20,
    xxl: 24,
    xxs: 4,
    "3xl": 32,
    none: 0,
    "3xs": 2,
    "4xl": 40,
    "5xl": 48,
    "6xl": 56,
    "7xl": 64,
    "8xl": 96,
    "9xl": 128,
  },
};